+++ NEWS.md
@@ -1,4 +1,4 @@
-- Added `ypr_yield_limits()` to account for uncertainty in parameters.
+- Added `ypr_yield_uncertainty()` to account for uncertainty in parameters.
 - Increased upper limit for `Rk` from 10 to 100.
 - Removed `a` and `fa` as just scaling terms.